Move `swap_digest.rs` under `digest` crate
